weather and emoji thumbs up/down

+++ src/main/java/org/wyvie/chehov/bot/commands/weather/model/Wind.java
@@ -3,7 +3,7 @@ package org.wyvie.chehov.bot.commands.weather.model;
 public class Wind {
 
     private double speed;
-    private int deg;
+    private double deg;
 
     public double getSpeed() {
         return speed;
@@ -13,11 +13,11 @@ public class Wind {
         this.speed = speed;
     }
 
-    public int getDeg() {
+    public double getDeg() {
         return deg;
     }
 
-    public void setDeg(int deg) {
+    public void setDeg(double deg) {
         this.deg = deg;
     }
 }
